changement message pour execution

+++ Serveur/ChatServeur.java
@@ -95,7 +95,7 @@ public class ChatServeur extends UnicastRemoteObject implements ChatGestion {
 	  int port; String URL;
 	  try { // transformation d une chaine de caracteres en entier
 		  //Integer I = new Integer(args[0]); port = I.intValue(); pour avoir en numero de port le paramètre du main
-		  Integer I = new Integer(1098); port = I.intValue();
+		  Integer I = new Integer(1097); port = I.intValue();
 	  } catch (Exception ex) {
 		  System.out.println(" Please enter: Server <port>"); return;
 	  }
